backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
google-generativeai==0.3.2
openai==1.12.0
python-multipart==0.0.6
pydantic-settings==2.1.0
pydantic[email]==2.5.3
python-dotenv==1.0.0
httpx==0.27.2
